avoid setting global slog
